Add files via upload
* You WON'T be able to keep records upgrading to this Version *
Upgrading to this version requires a uninstall of the plugin entirely and then reinstallation of this new version.

New database structure brings it inline with MyBB standard which also allows more people using an older database to use this plugin.
New filtering options allow you to limit how many results are returned which will be immensely useful in larger forums and also includes user filtering and sorting options which I expect will be very helpful. 

You can of course still filter using the DataTable by clicking on the columns and searching in there as well. This update is quite big in comparison to code changed but I hope you all enjoy it.

Let me know if there's any more suggestions/errors. And as usual for more updates/news on my developments follow me on twitter: https://twitter.com/_JCrookshank

Cheers